+++ pcap4j-core/src/main/java/org/pcap4j/packet/IcmpV6NeighborSolicitationPacket.java
@@ -247,7 +247,7 @@ public final class IcmpV6NeighborSolicitationPacket extends AbstractPacket {
     }
 
     @Override
-    protected int measureLength() {
+    protected int calcLength() {
       int len = 0;
       for (IpV6NeighborDiscoveryOption o: options) {
         len += o.length();