+++ pcap4j-core/src/main/java/org/pcap4j/packet/IcmpV6RedirectPacket.java
@@ -287,7 +287,7 @@ public final class IcmpV6RedirectPacket extends AbstractPacket {
     }
 
     @Override
-    protected int measureLength() {
+    protected int calcLength() {
       int len = 0;
       for (IpV6NeighborDiscoveryOption o: options) {
         len += o.length();