+++ pcap4j-core/src/main/java/org/pcap4j/packet/IpV4Packet.java
@@ -836,7 +836,7 @@ public final class IpV4Packet extends AbstractPacket {
     }
 
     @Override
-    protected int measureLength() {
+    protected int calcLength() {
       return measureLengthWithoutPadding() + padding.length;
     }